refactor: 장소 검색에 관한 type을 한 파일로 관리
- SearchEmptyState -> SearchEmptyList로 변경
