--- frontend/src/domains/places/components/SearchAddress/SearchAddress.types.ts
@@ -1,6 +1,0 @@
-interface SearchAddressProps {
-  addressType: '도로명' | '지번';
-  address: string;
-}
-
-export type { SearchAddressProps };
