--- frontend/src/domains/places/components/SearchBox/SearchBox.types.ts
@@ -1,5 +1,0 @@
-interface SearchBoxProps {
-  onClose: () => void;
-}
-
-export type { SearchBoxProps };
