--- frontend/src/domains/places/components/SearchList/SearchList.types.ts
@@ -1,13 +1,0 @@
-import type { SearchedPlaceType } from '../../types/place.types';
-
-interface SearchEmptyStateProps {
-  keyword: string;
-}
-
-interface SearchListProps {
-  searchResults: SearchedPlaceType[];
-  onClose: () => void;
-  searchedKeyword: string;
-}
-
-export type { SearchEmptyStateProps, SearchListProps };
