+++ frontend/src/domains/places/types/searchPlace.types.ts
@@ -1,0 +1,36 @@
+import { SearchedPlaceType } from './place.types';
+
+interface SearchBoxProps {
+  onClose: () => void;
+}
+
+interface SearchEmptyListProps {
+  keyword: string;
+}
+
+interface SearchListProps {
+  searchResults: SearchedPlaceType[];
+  onClose: () => void;
+  searchedKeyword: string;
+}
+
+type AddressType = '도로명' | '지번';
+
+interface SearchAddressProps {
+  addressType: AddressType;
+  address: string;
+}
+
+interface SearchPlaceProps extends SearchAddressProps {
+  searchResult: SearchedPlaceType;
+  onClose: () => void;
+}
+
+export type {
+  SearchBoxProps,
+  SearchEmptyListProps,
+  SearchListProps,
+  AddressType,
+  SearchAddressProps,
+  SearchPlaceProps,
+};
